requirements.txt
amqp==5.1.1
asgiref==3.6.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.12
django-bootstrap3==23.1
django-celery-results==2.5.1
django-filter==23.2
django-tables2==2.5.3
kombu==5.2.4
prompt-toolkit==3.0.38
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
vine==5.0.0
wcwidth==0.2.6
